feat: welcome page is root page, settings and chat protected routes
