--- nextjs/src/app/dashboard/page.tsx
@@ -1,5 +1,0 @@
-const Dashboard = () => {
-    return <p>{"You're signed in! Welcome to the dashboard."}</p>;
-};
-
-export default Dashboard;
